week-4/day-5/Mini-Project/pyproject.toml
[project]
name = "mini-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "joblib>=1.5.3",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "streamlit>=1.54.0",
]
